x/blog/keeper/grpc_query_bridge_status.go
package keeper

import (
	"context"

	sdk "github.com/cosmos/cosmos-sdk/types"
	"google.golang.org/grpc/codes"
	"google.golang.org/grpc/status"
	"planet/x/blog/types"
)

func (k Keeper) BridgeStatus(c context.Context, req *types.QueryGetBridgeStatusRequest) (*types.QueryGetBridgeStatusResponse, error) {
	if req == nil {
		return nil, status.Error(codes.InvalidArgument, "invalid request")
	}
	ctx := sdk.UnwrapSDKContext(c)

	val, found := k.GetBridgeStatus(ctx)
	if !found {
		return nil, status.Error(codes.NotFound, "not found")
	}

	return &types.QueryGetBridgeStatusResponse{BridgeStatus: val}, nil
}
